requirements.txt
alembic==1.7.5
blinker==1.4
Click==7.0
dnspython==2.1.0
email-validator==1.1.3
Flask==1.1.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==1.0.0
gunicorn==20.0.4
idna==3.3
importlib-metadata==4.8.0
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.6
MarkupSafe==1.1.0
PyJWT==2.3.0
python-dotenv==0.19.2
SQLAlchemy==1.3.12
Werkzeug==0.16.1
WTForms==3.0.0
zipp==3.6.0
